pyproject.toml
[project]
name = "py-doc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "docling>=2.31.1",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "transformers>=4.42.4",
    "timm>=1.0.15",
    "pymupdf>=1.25.5",
    "ocrmac>=1.0.0",
    "ultralytics>=8.3.133",
    "tesseract>=0.1.3",
    "tessdata>=1.0.0",
    "python-docx>=1.2.0",
    "docx2pdf>=0.1.8",
    "flask>=3.1.1",
    "openai>=1.97.1",
]

[dependency-groups]
dev = [
    "ruff>=0.13.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
